requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
streamlit>=1.31.0
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.5.0
typing-extensions>=4.8.0
groq>=0.4.0
pandas>=2.0.0
plotly>=5.18.0
numpy<2.0.0 